refactor: replace model in types
