--- src/models/teamMemberModel.ts
@@ -1,7 +1,0 @@
-export interface TeamMemberModel {
-    id: string,
-    nickname: string;
-    avatar: string;
-    name: string;
-    description: string;
-}
