+++ src/types/index.ts
@@ -86,3 +86,11 @@ export interface EvaluationTaskModel {
   score: number;
   done: boolean;
 }
+
+export interface TeamMemberModel {
+  id: string;
+  nickname: string;
+  avatar: string;
+  name: string;
+  description: string;
+}
